minimum price threshold

+++ analyst/__main__.py
@@ -16,7 +16,7 @@ parser_screener.set_defaults(fn=run_screener_task)
 
 parser_get_stock_data = subparsers.add_parser("getstockdata")
 parser_get_stock_data.add_argument(
-    "minimum_price", help="The minimum price threshold.", type=int
+    "minimum_price", help="The minimum price threshold.", type=float
 )
 parser_get_stock_data.set_defaults(fn=run_get_stock_data_task)
